% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.all.R
\name{mda.all}
\alias{mda.all}
\title{Return all multidiffabundance tools or a selection on an mda object.}
\usage{
mda.all(
  mda.D,
  alpha = FALSE,
  beta = FALSE,
  group = FALSE,
  continuous = FALSE,
  aldex2 = TRUE,
  ancombc = TRUE,
  corncob = FALSE,
  deseq2 = TRUE,
  limma = TRUE,
  lmclr = TRUE,
  maaslin2 = TRUE,
  zicoseq = TRUE,
  ...
)
}
\arguments{
\item{alpha}{if an alpha diversity test is to be performed. Defaults to false.}

\item{beta}{if a beta diversity test is to be performed. Defaults to false.}

\item{group}{if a grouped analysis is to be performed. Defaults to false.}

\item{continuous}{if a continuous value analysis is to be performed. Defaults to false.}

\item{aldex2}{if the aldex2 mda pipeline is to be performed. Defaults to true.}

\item{ancombc}{if the ancombc mda pipeline is to be performed. Defaults to true.}

\item{corncob}{if the corncob mda pipeline is to be performed. Defaults to false, due to long runtime of the analysis.}

\item{deseq2}{if the deseq2 mda pipeline is to be performed. Defaults to true.}

\item{limma}{if the limma mda pipeline is to be performed. Defaults to true.}

\item{lmclr}{if the lmclr mda pipeline is to be performed. Defaults to true.}

\item{maaslin2}{if the maaslin2 mda pipeline is to be performed. Defaults to true.}

\item{zicoseq}{if the zicoseq mda pipeline is to be performed. Defaults to true.}

\item{D}{An mda object generated by \code{\link{mda.create}}.}

\item{group.cols}{column names of the group values to be analyzed. Required if group is true.}

\item{continuous.cols}{column names of the continuous values to be analyzed. Required if continuous is true.}

\item{continuous.scale}{if the continuous values should be scaled. Defaults to true.}
}
\description{
Return all multidiffabundance tools or a selection on an mda object.
}
