% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.continuous.R
\name{mda.continuous}
\alias{mda.continuous}
\title{Continuous value analysis}
\usage{
mda.continuous(mda.D, continuous.cols = NULL, continuous.scale = TRUE, ...)
}
\arguments{
\item{continuous.cols}{list of columns with continuous variables}

\item{continuous.scale}{wether to scale the variables or not}
}
\description{
Continuous value analysis
}
