% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{mda.from_tidytacos}
\alias{mda.from_tidytacos}
\title{Create an mda object from a tidytacos object and a formula}
\usage{
mda.from_tidytacos(ta, formulas, output_dir = tempdir(), ...)
}
\description{
Create an mda object from a tidytacos object and a formula
}
