% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.limma.R
\name{mda.limma}
\alias{mda.limma}
\title{Limma Differential Abundance analysis}
\usage{
mda.limma(mda.D, ...)
}
\arguments{
\item{mda.D}{MDA object}

\item{...}{
  Arguments passed on to \code{\link[limma:ebayes]{limma::eBayes}}
  \describe{
    \item{\code{fit}}{an \code{MArrayLM} fitted model object produced by \code{lmFit} or \code{contrasts.fit}.
  For \code{ebayes} only, \code{fit} can alternatively be an unclassed list produced by \code{lm.series}, \code{gls.series} or \code{mrlm} containing components \code{coefficients}, \code{stdev.unscaled}, \code{sigma} and \code{df.residual}.}
    \item{\code{proportion}}{numeric value between 0 and 1, assumed proportion of genes which are differentially expressed}
    \item{\code{stdev.coef.lim}}{numeric vector of length 2, assumed lower and upper limits for the standard deviation of log2-fold-changes for differentially expressed genes}
    \item{\code{trend}}{logical, should an intensity-dependent trend be allowed for the prior variance? If \code{FALSE} then the prior variance is constant. Alternatively, \code{trend} can be a row-wise numeric vector, which will be used as the covariate for the prior variance.}
    \item{\code{robust}}{logical, should the estimation of \code{df.prior} and \code{var.prior} be robustified against outlier sample variances?}
    \item{\code{winsor.tail.p}}{numeric vector of length 1 or 2, giving left and right tail proportions of \code{x} to Winsorize (if \code{robust=TRUE} and \code{legacy=TRUE}).}
    \item{\code{legacy}}{logical. If \code{FALSE} then the new hyperparameter estimation (introduced in limma 3.61.8) will be used, if \code{TRUE} the earlier hyperparameter estimation will be used. If \code{NULL}, then the new method will be used when the residual degrees of freedom are not all equal and the old method will be used otherwise.}
  }}
}
\description{
Limma Differential Abundance analysis
}
