% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{mda.merge_results}
\alias{mda.merge_results}
\title{A function to merge the outputs of several mda runs}
\usage{
mda.merge_results(res.list)
}
\description{
A function to merge the outputs of several mda runs
}
