% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{mda.permute_formula}
\alias{mda.permute_formula}
\title{Permutes over all possible orders of effects in a given formula and returns a list of the possibilities.}
\usage{
mda.permute_formula(form)
}
\arguments{
\item{form}{a formula to permute over, eg: ~ a:b + a + b.}
}
\description{
Permutes over all possible orders of effects in a given formula and returns a list of the possibilities.
}
