% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.plots.R
\name{mda.plot.all}
\alias{mda.plot.all}
\title{Return a heatplot, beta plot, and alpha plot in a single plotly object.}
\usage{
mda.plot.all(res, ...)
}
\arguments{
\item{res}{The result of an mda.all(D, alpha=TRUE) or mda.alpha run.}

\item{low}{The color to use for low effect sizes.}

\item{high}{The color to use for high effect sizes.}

\item{mid}{The color to use for effect sizes around 0.}
}
\description{
Return a heatplot, beta plot, and alpha plot in a single plotly object.
}
