% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.plots.R
\name{mda.plot.alpha}
\alias{mda.plot.alpha}
\title{Return an interactive plot of the effect sizes of the covariates on the alpha diversity}
\usage{
mda.plot.alpha(res)
}
\arguments{
\item{res}{The result of an mda.all(D, alpha=TRUE) or mda.alpha run.}
}
\description{
Return an interactive plot of the effect sizes of the covariates on the alpha diversity
}
