% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.plots.R
\name{mda.plot.heatplot}
\alias{mda.plot.heatplot}
\title{Return an interactive differential abundance heatplot of the effect sizes of the covariates on the taxa using different methods}
\usage{
mda.plot.heatplot(
  res,
  ta = NULL,
  low = "magenta",
  high = "limegreen",
  mid = "white"
)
}
\arguments{
\item{res}{The result of an mda.all(D, alpha=TRUE) or mda.alpha run.}

\item{low}{The color to use for low effect sizes.}

\item{high}{The color to use for high effect sizes.}

\item{mid}{The color to use for effect sizes around 0.}
}
\description{
Return an interactive differential abundance heatplot of the effect sizes of the covariates on the taxa using different methods
}
