% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{mda.summary}
\alias{mda.summary}
\title{Summary function}
\usage{
mda.summary(
  res,
  id_cols = "taxa",
  names_from = "variable",
  method_from = "method",
  pvalue_from = "pvalue",
  qvalue_from = "qvalue",
  effectsize_from = "effectsize",
  values_fn = list,
  pvalue_threshold = 0.05,
  qvalue_threshold = 0.05,
  method.groups = list(da = c("aldex2", "ancombc2", "corncob", "deseq2", "corncob",
    "limma", "lmclr", "maaslin2", "zicoseq"), alpha = c("alpha"), beta = c("beta"), group
    = c("group"), continuous = c("continuous"))
)
}
\description{
Summary function
}
