% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.write.R
\name{mda.write}
\alias{mda.write}
\title{Write an mda object to two tsv tables and a file containing the formulas in a folder}
\usage{
mda.write(mda, folder = "mda")
}
\arguments{
\item{mda}{the mda object}

\item{folder}{the folder to save the tables and formulas to}
}
\description{
Write an mda object to two tsv tables and a file containing the formulas in a folder
}
